// src/shared/Brackets/types.tsx
export type Options = {
    width?: number;

    boxHeight?: number;

    canvasPadding?: number;

    spaceBetweenColumns?: number;

    spaceBetweenRows?: number;

    connectorColor?: string;

    connectorColorHighlight?: string;

    roundHeader?: {
        isShown?: boolean;
        height?: number;
        marginBottom?: number;
        fontSize?: number;
        fontColor?: string;
        backgroundColor?: string;
        fontFamily?: string;
        roundTextGenerator?: (
            currentRoundNumber: number,
            roundsTotalNumber: number
        ) => string | undefined;
    };

    roundSeparatorWidth?: number;

    lineInfo?: {
        separation?: number;
        homeVisitorSpread?: number;
    };

    horizontalOffset?: number;

    wonBywalkOverText?: string;

    lostByNoShowText?: string;
};

export type ComputedOptions = Options & {
    rowHeight?: number;

    columnWidth?: number;
};